Move command into cumulus CLI
