+++ cumulus/scripts/__init__.py
@@ -1,0 +1,1 @@
+from command import cli
